+++ src/stages/playable_stage.cpp
@@ -8,23 +8,42 @@
 
 namespace artifact
 {
-    PlayableStage::PlayableStage(const char *identifier) : Stage(identifier) {}
+    PlayableStage::PlayableStage(const char *identifier) :
+        Stage(identifier) {}
     void PlayableStage::startup()
     {
         Stage::startup();
-        // Spawn the Player
-        this->player = spawn_entity<PlayerEntity>(0, 0);
+        spawn_entities();
+
+        level_open_overlay = new LevelOpenOverlay(this, 1.f);
+        level_open_overlay->pause(); // Don't start animation until first update
+        pause_screen = std::make_unique<PauseScreen>(this);
+        death_screen = std::make_unique<DeathScreen>(this);
     }
     void PlayableStage::draw() const
     {
-
+        if (is_being_destroyed)
+            return;
         Stage::draw();
+
         DrawTextureEx(background, {-1000, -3150}, 0, 2.3, WHITE);
         if (Game::get_instance()->debug_mode)
             debug_draw_colliders();
         for (auto &entity: entities)
         {
             entity->draw();
+        }
+    }
+    void PlayableStage::draw_ui() const
+    {
+        this->player->draw_stats();
+        level_open_overlay->draw();
+
+        if (player->is_dead())
+            death_screen->draw();
+        if (is_paused)
+        {
+            pause_screen->draw();
         }
     }
     void PlayableStage::debug_draw_colliders() const
@@ -34,30 +53,96 @@ namespace artifact
             DrawRectangleLinesEx(collider.bounds, 4, collider.is_blocking ? RED : BLUE);
         }
     }
-    void PlayableStage::update(const float deltaTime)
+    void PlayableStage::update(const float delta_time)
     {
-        Stage::update(deltaTime);
+        if (is_being_destroyed)
+            return;
+        Stage::update(delta_time);
+
+        // Start animation on first update frame after stage is fully loaded
+        if (is_first_frame)
+        {
+            level_open_overlay->play();
+            is_first_frame = false;
+        }
+
+        level_open_overlay->update(delta_time);
+        if (IsKeyPressed(KEY_ESCAPE))
+            is_paused = !is_paused;
+        if (is_paused)
+        {
+            if (this->peek_zindex() != pause_screen.get())
+                this->push_to_zindex(pause_screen.get());
+            return;
+        }
+
+        if (this->peek_zindex() == pause_screen.get())
+            this->remove_from_zindex(pause_screen.get());
+
 
         for (const auto &entity: entities)
         {
-            entity->update(deltaTime);
+            entity->update(delta_time);
+            entity->check_entity_collisions(entities);
         }
+
+        check_collider_overlaps();
+    }
+    void PlayableStage::update(const int mouse_x, const int mouse_y)
+    {
+        Stage::update(mouse_x, mouse_y);
+        if (player->is_dead())
+            death_screen->update(mouse_x, mouse_y);
+        if (is_paused)
+            pause_screen->update(mouse_x, mouse_y);
     }
     void PlayableStage::destroy()
     {
+        if (is_being_destroyed)
+            return;
         Stage::destroy();
         entities.clear();
         UnloadTexture(background);
     }
-    bool PlayableStage::is_entity_colliding(const Entity *entity) const { return Collider::is_entity_colliding(entity, colliders); }
     Collider PlayableStage::get_collider_at(const int x, const int y, const bool blocking_only) const { return Collider::get_collider_at(x, y, colliders, blocking_only); }
     std::vector<Collider> PlayableStage::get_colliders_closest_to(const int x, const int y, const bool blocking_only) const { return Collider::get_colliders_closest_to(x, y, colliders, blocking_only); }
     std::vector<Collider> PlayableStage::get_blocking_colliders() const { return Collider::get_blocking_colliders(colliders); }
     void PlayableStage::set_background(const char *resource_location) { this->background = LoadTexture(resource_location); }
     Texture2D *PlayableStage::get_background() { return &background; }
+    void PlayableStage::pause() { is_paused = true; }
+    void PlayableStage::unpause() { is_paused = false; }
+    Vector2 PlayableStage::get_spawn_position() const { return {0, 0}; }
     void PlayableStage::respawn()
     {
-        camera.target = {0, 0};
-        player->set_position(0, 0);
+        player->respawn();
     }
+    PlayerEntity *PlayableStage::get_player()
+    {
+        return player;
+    }
+
+    void PlayableStage::check_collider_overlaps()
+    {
+        for (const auto &entity: entities)
+        {
+            const Rectangle entityBounds = {
+                    entity->get_position().x,
+                    entity->get_position().y,
+                    entity->get_width(),
+                    entity->get_height()};
+
+            for (auto &collider: colliders)
+            {
+                if (!collider.is_blocking && CheckCollisionRecs(entityBounds, collider.bounds))
+                {
+                    collider.overlap(entity);
+                }
+            }
+        }
+    }
+    void PlayableStage::register_collider(const int x, const int y, const int width, const int height, const std::function<void(Entity *)> &on_entity_overlap)
+    {
+        colliders.emplace_back(x, y, width, height, on_entity_overlap);
+    }
+
 } // namespace artifact
